+++ cqp/latex-print.h
@@ -41,12 +41,12 @@ latex_print_corpus_header(CorpusList *cl,
 void 
 latex_print_output(CorpusList *cl, 
 		   FILE *outfd,
-		   int interactive,
+		   int64_t interactive,
 		   ContextDescriptor *cd,
-		   int first, int last);
+		   int64_t first, int64_t last);
 
 void 
-latex_print_group(Group *group, int expand, FILE *fd);
+latex_print_group(Group *group, int64_t expand, FILE *fd);
 
 
 #endif
